+++ flask_restplus/__about__.py
@@ -1,3 +1,3 @@
 # -*- coding: utf-8 -*-
-__version__ = '0.6.1.dev'
+__version__ = '0.7.0'
 __description__ = 'Helpers, syntaxic sugar and Swagger documentation for Flask-Restful'