+++ CHANGELOG.rst
@@ -7,6 +7,8 @@ Current
 - Expose models and fields schema through the ``__schema__`` attribute
 - Drop support for model as class
 - Added ``@api.errorhandler()`` to register custom error handlers
+- Added ``@api.response''`` shortcut decorator
+- Fix list nested models missing in definitions
 
 
 0.6.0